:recycle: Refactor ErrorMessage to SupportOrErrorMessage
