+++ src/components/CheckBox/index.tsx
@@ -1,7 +1,7 @@
 import { cva, type VariantProps } from "class-variance-authority";
 import { HTMLProps } from "react";
-import ErrorMessage from "src/internal/ErrorMessage";
 import FormLabel from "src/internal/FormLabel";
+import SupportOrErrorMessage from "src/internal/SupportOrErrorMessage";
 import { twMerge } from "tailwind-merge";
 
 export interface ICheckBox {
@@ -139,7 +139,7 @@ export const CheckBox = ({
           </svg>
         </label>
       </div>
-      <ErrorMessage error={error} errorMsg={errorMsg} supportText={supportText} />
+      <SupportOrErrorMessage error={error} errorMsg={errorMsg} supportText={supportText} />
     </div>
   );
 };
