+++ src/components/SelectBox/index.tsx
@@ -10,7 +10,7 @@ import {
 import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from "@heroicons/react/24/outline";
 import { cva, type VariantProps } from "class-variance-authority";
 import { ElementType, Fragment, ReactNode, useEffect, useState } from "react";
-import ErrorMessage from "src/internal/ErrorMessage";
+import SupportOrErrorMessage from "src/internal/SupportOrErrorMessage";
 import { twMerge } from "tailwind-merge";
 
 type Options = {
@@ -221,7 +221,7 @@ function SelectBox({
           </>
         )}
       </Listbox>
-      <ErrorMessage error={error} errorMsg={errorMsg} supportText={supportText} />
+      <SupportOrErrorMessage error={error} errorMsg={errorMsg} supportText={supportText} />
     </div>
   );
 }
