+++ src/components/TextField/TextFieldBase.tsx
@@ -1,7 +1,7 @@
 import { cva, VariantProps } from "class-variance-authority";
 import React, { ElementType, forwardRef } from "react";
 import InputMask from "react-input-mask";
-import ErrorMessage from "src/internal/ErrorMessage";
+import SupportOrErrorMessage from "src/internal/SupportOrErrorMessage";
 import { twMerge } from "tailwind-merge";
 
 import FormLabel from "../../internal/FormLabel";
@@ -148,7 +148,7 @@ export const TextFieldBase = /* @__PURE__ */ forwardRef(
             />
           )}
         </div>
-        <ErrorMessage error={error} errorMsg={errorMsg} supportText={supportText} />
+        <SupportOrErrorMessage error={error} errorMsg={errorMsg} supportText={supportText} />
       </div>
     );
   },
